+++ milvus/client/http_handler.py
@@ -1,17 +1,17 @@
-import requests as rq
-import json
-import time
-import datetime
 import copy
-from urllib.parse import urlparse
+import datetime
+import json
 import logging
+import requests as rq
+import struct
+import time
+from urllib.parse import urlparse
 
 from .abstract import ConnectIntf, IndexParam, TableSchema, TopKQueryResult2, PartitionParam
-from .types import Status, IndexType, MetricType
-from .exceptions import ParamError
 from .check import is_legal_host, is_legal_port
-from .exceptions import NotConnectError
+from .exceptions import ParamError, NotConnectError
 from .prepare import *
+from .types import Status, IndexType, MetricType
 
 from ..settings import DefaultConfig as config
 
@@ -23,6 +23,7 @@ IndexValue2NameMap = {
     IndexType.FLAT: "FLAT",
     IndexType.IVFLAT: "IVFFLAT",
     IndexType.IVF_SQ8: "IVFSQ8",
+    IndexType.IVF_SQ8H: "IVFSQ8H",
     IndexType.IVF_PQ: "IVFPQ"
 }
 
@@ -31,6 +32,7 @@ IndexName2ValueMap = {
     "FLAT": IndexType.FLAT,
     "IVFFLAT": IndexType.IVFLAT,
     "IVFSQ8": IndexType.IVF_SQ8,
+    "IVFSQ8H": IndexType.IVF_SQ8H,
     "IVFPQ": IndexType.IVF_PQ
 }
 
@@ -123,10 +125,11 @@ class HttpHandler(ConnectIntf):
         return self._status
 
     def connected(self):
-        return self.ping()
+        return False if self._status is None else self.ping()
 
     def disconnect(self):
-        pass
+        self._status = None
+        return Status()
 
     def create_table(self, param, timeout):
 
@@ -148,7 +151,6 @@ class HttpHandler(ConnectIntf):
             return Status(Status.UNEXPECTED_ERROR, message=str(e))
 
     def has_table(self, table_name, timeout):
-        _ = Prepare.table_name(table_name)
 
         url = self._uri + "/tables/" + table_name
         try:
@@ -166,8 +168,6 @@ class HttpHandler(ConnectIntf):
             return Status(Status.UNEXPECTED_ERROR, message=str(e))
 
     def count_table(self, table_name, timeout):
-        _ = Prepare.table_name(table_name)
-
         url = self._uri + "/tables/{}".format(table_name)
 
         try:
@@ -184,8 +184,6 @@ class HttpHandler(ConnectIntf):
             return Status(Status.UNEXPECTED_ERROR, message=str(e)), None
 
     def describe_table(self, table_name, timeout):
-        _ = Prepare.table_name(table_name)
-
         url = self._uri + "/tables/{}".format(table_name)
 
         try:
@@ -256,11 +254,7 @@ class HttpHandler(ConnectIntf):
             return Status(Status.UNEXPECTED_ERROR, message=str(e))
 
     def insert(self, table_name, records, ids, partition_tag, timeout, **kwargs):
-        _ = Prepare.insert_param(table_name, records, partition_tag, ids)
-
         url = self._uri + "/tables/{}/vectors".format(table_name)
-
-        t0 = time.time()
 
         data_dict = dict()
         if ids:
@@ -268,35 +262,19 @@ class HttpHandler(ConnectIntf):
         if partition_tag:
             data_dict["tag"] = partition_tag
 
-        data_dict["records"] = records
-
-        t1 = time.time()
+        if isinstance(records[0], bytes):
+            vectors = [struct.unpack(str(len(r)) + 'B', r) for r in records]
+        else:
+            vectors = records
+        data_dict["records"] = vectors
 
         data = json.dumps(data_dict)
-
-        t2 = time.time()
 
         headers = {"Content-Type": "application/json"}
 
         try:
-            time_stamp1 = datetime.datetime.now()
             response = rq.post(url, data=data, headers=headers)
-
-            time_stamp2 = datetime.datetime.now()
-
-            print("[{}] before request".format(time_stamp1))
-            print("[{}] after request".format(time_stamp2))
-
-            t3 = time.time()
-
             js = response.json()
-
-            t4 = time.time()
-
-            print("Add records to dict cost {:.4f} s".format(t1 - t0))
-            print("Json dumps cost {:.4f} s".format(t2 - t1))
-            print("Request cost {:.4f} s".format(t3 - t2))
-            print("Response json cost {:.4f} s".format(t4 - t3))
 
             if 201 == response.status_code:
                 ids = [int(item) for item in list(js["ids"])]
@@ -308,10 +286,7 @@ class HttpHandler(ConnectIntf):
         except Exception as e:
             return Status(Status.UNEXPECTED_ERROR, message=str(e)), []
 
-    def search(self, table_name, top_k, nprobe,
-               query_records, query_ranges=None, partition_tags=None, **kwargs):
-
-        _ = Prepare.search_param(table_name, top_k, nprobe, query_records, query_ranges, partition_tags)
+    def search(self, table_name, top_k, nprobe, query_records, partition_tags=None, **kwargs):
 
         url = self._uri + "/tables/{}/vectors".format(table_name)
 
@@ -320,7 +295,13 @@ class HttpHandler(ConnectIntf):
             body_dict["tags"] = partition_tags
         body_dict["topk"] = top_k
         body_dict["nprobe"] = nprobe
-        body_dict["records"] = query_records
+
+        if isinstance(query_records[0], bytes):
+            vectors = [struct.unpack(str(len(r)) + 'B', r) for r in query_records]
+        else:
+            vectors = query_records
+
+        body_dict["records"] = vectors
         data = json.dumps(body_dict)
         headers = {"Content-Type": "application/json"}
 
@@ -337,16 +318,20 @@ class HttpHandler(ConnectIntf):
         except Exception as e:
             return Status(Status.UNEXPECTED_ERROR, message=str(e)), None
 
-    def search_in_files(self, table_name, file_ids, query_records, top_k,
-                        nprobe=16, query_ranges=None, **kwargs):
-        _ = Prepare.search_vector_in_files_param(table_name, query_records, query_ranges, top_k, nprobe, file_ids)
+    def search_in_files(self, table_name, file_ids, query_records, top_k, nprobe=16, **kwargs):
         url = self._uri + "/tables/{}/vectors".format(table_name)
 
         body_dict = dict()
         body_dict["topk"] = top_k
         body_dict["nprobe"] = nprobe
-        body_dict["records"] = query_records
         body_dict["file_ids"] = file_ids
+
+        if isinstance(query_records[0], bytes):
+            vectors = [struct.unpack(str(len(r)) + 'B', r) for r in query_records]
+        else:
+            vectors = query_records
+        body_dict["records"] = vectors
+
         data = json.dumps(body_dict)
         headers = {"Content-Type": "application/json"}
 
@@ -363,8 +348,11 @@ class HttpHandler(ConnectIntf):
         except Exception as e:
             return Status(Status.UNEXPECTED_ERROR, message=str(e)), None
 
+    def search_by_id(self, table_name, top_k, nprobe, id_array, partition_tag_array):
+
+        return Status(Status.UNEXPECTED_ERROR, "Not uncompleted"), None
+
     def create_index(self, table_name, index, timeout):
-        _ = Prepare.index_param(table_name, index)
 
         try:
             url = self._uri + "/tables/{}/indexes".format(table_name)
@@ -391,10 +379,8 @@ class HttpHandler(ConnectIntf):
 
     def preload_table(self, table_name, timeout):
         return Status()
-        # super().preload_table(table_name, timeout)
 
     def describe_index(self, table_name, timeout):
-        _ = Prepare.table_name(table_name)
         url = self._uri + "/tables/{}/indexes".format(table_name)
 
         try:
@@ -418,7 +404,6 @@ class HttpHandler(ConnectIntf):
             return Status(Status.UNEXPECTED_ERROR, message=str(e))
 
     def drop_index(self, table_name, timeout):
-        _ = Prepare.table_name(table_name)
 
         try:
             url = self._uri + "/tables/{}/indexes".format(table_name)
@@ -452,7 +437,6 @@ class HttpHandler(ConnectIntf):
             return Status(Status.UNEXPECTED_ERROR, message=str(e)), None
 
     def create_partition(self, table_name, partition_name, partition_tag, timeout=10):
-        _ = Prepare.partition_param(table_name, partition_name, partition_tag)
 
         url = self._uri + "/tables/{}/partitions".format(table_name)
 
@@ -472,7 +456,6 @@ class HttpHandler(ConnectIntf):
             return Status(Status.UNEXPECTED_ERROR, message=str(e))
 
     def show_partitions(self, table_name, timeout):
-        _ = Prepare.table_name(table_name)
 
         url = self._uri + "/tables/{}/partitions".format(table_name)
         query_data = {"offset": 0, "page_size": 100}
@@ -497,6 +480,7 @@ class HttpHandler(ConnectIntf):
             return Status(Status.UNEXPECTED_ERROR, message=str(e)), []
 
     def drop_partition(self, table_name, partition_tag, timeout=10):
+
         url = self._uri + "/tables/{}/partitions/{}".format(table_name, partition_tag)
 
         try:
@@ -510,3 +494,9 @@ class HttpHandler(ConnectIntf):
             return Status(Status.UNEXPECTED_ERROR, message='Request timeout')
         except Exception as e:
             return Status(Status.UNEXPECTED_ERROR, message=str(e))
+
+    def delete_by_id(self, table_name, id_array, timeout=None):
+        return Status()
+
+    def flush(self, table_name_array):
+        return Status()